pyproject.toml
[project]
name = "anklume"
version = "0.1.0"
description = "QubesOS-like isolation using native Linux kernel features, orchestrated by Ansible and Incus"
requires-python = ">=3.11"
dependencies = ["typer[all]>=0.12", "pyyaml>=6.0", "fastapi>=0.115", "uvicorn>=0.32"]

[project.scripts]
anklume = "scripts.cli:app"

[project.optional-dependencies]
test = ["pytest", "pytest-xdist", "pyyaml", "hypothesis", "behave"]
docs = ["mkdocs-material", "mkdocs-static-i18n>=1.2"]
live-test = [
    "vncdotool>=1.2",
    "opencv-python-headless>=4.8",
    "pytesseract>=0.3.10",
    "pexpect>=4.9",
    "pillow>=10.0",
    "pixelmatch>=0.3",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
